chore(design-system): move MessageItem UI components from design system to :feature:mail:message:list:api
