+++ feature/mail/message/list/api/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(projects.core.logging.api)
     implementation(projects.core.preference.api)
     implementation(projects.core.ui.compose.common)
+    implementation(projects.core.ui.compose.designsystem)
     implementation(projects.feature.account.api)
     implementation(projects.feature.mail.account.api)
     implementation(projects.feature.mail.folder.api)
