+++ feature/mail/message/list/api/src/debug/kotlin/net/thunderbird/feature/mail/message/list/ui/component/organism/JunkMessageItemPreview.kt
@@ -1,4 +1,4 @@
-package net.thunderbird.core.ui.compose.designsystem.organism.message
+package net.thunderbird.feature.mail.message.list.ui.component.organism
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -19,7 +19,7 @@ import app.k9mail.core.ui.compose.designsystem.PreviewWithThemes
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextTitleSmall
 import app.k9mail.core.ui.compose.theme2.MainTheme
 
-private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<MessageItemPrevParams>(
+private class JunkMessageItemPrevParamCol : CollectionPreviewParameterProvider<MessageItemPrevParams>(
     collection = listOf(
         MessageItemPrevParams(
             sender = "Sender Name",
@@ -57,7 +57,6 @@ private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<M
             preview = LoremIpsum(words = 20).values.joinToString(),
             hasAttachments = true,
             selected = true,
-            favourite = true,
             threadCount = 100,
             swapSenderWithSubject = false,
         ),
@@ -67,7 +66,6 @@ private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<M
             preview = LoremIpsum(words = 3).values.joinToString(),
             hasAttachments = false,
             selected = false,
-            favourite = false,
             threadCount = 0,
             swapSenderWithSubject = true,
         ),
@@ -77,7 +75,6 @@ private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<M
             preview = LoremIpsum(words = 5).values.joinToString(),
             hasAttachments = true,
             selected = false,
-            favourite = false,
             threadCount = 1,
             swapSenderWithSubject = true,
         ),
@@ -87,7 +84,6 @@ private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<M
             preview = LoremIpsum(words = 10).values.joinToString(),
             hasAttachments = false,
             selected = true,
-            favourite = true,
             threadCount = 10,
             swapSenderWithSubject = true,
         ),
@@ -97,7 +93,6 @@ private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<M
             preview = LoremIpsum(words = 20).values.joinToString(),
             hasAttachments = true,
             selected = true,
-            favourite = true,
             threadCount = 100,
             swapSenderWithSubject = true,
         ),
@@ -107,10 +102,10 @@ private class ReadMessageItemPrevParamCol : CollectionPreviewParameterProvider<M
 @Preview
 @Composable
 private fun PreviewDefault(
-    @PreviewParameter(ReadMessageItemPrevParamCol::class) params: MessageItemPrevParams,
+    @PreviewParameter(JunkMessageItemPrevParamCol::class) params: MessageItemPrevParams,
 ) {
     PreviewWithThemes {
-        ReadMessageItem(
+        JunkMessageItem(
             sender = params.sender,
             subject = params.subject,
             preview = params.preview,
@@ -131,11 +126,9 @@ private fun PreviewDefault(
             onClick = { },
             onLongClick = { },
             onLeadingClick = { },
-            onFavouriteChange = { },
             modifier = Modifier.padding(MainTheme.spacings.double),
             hasAttachments = params.hasAttachments,
             selected = params.selected,
-            favourite = params.favourite,
             threadCount = params.threadCount,
             swapSenderWithSubject = params.swapSenderWithSubject,
             showAccountIndicator = true,
@@ -147,10 +140,10 @@ private fun PreviewDefault(
 @Preview
 @Composable
 private fun PreviewCompact(
-    @PreviewParameter(ReadMessageItemPrevParamCol::class) params: MessageItemPrevParams,
+    @PreviewParameter(JunkMessageItemPrevParamCol::class) params: MessageItemPrevParams,
 ) {
     PreviewWithThemes {
-        ReadMessageItem(
+        JunkMessageItem(
             sender = params.sender,
             subject = params.subject,
             preview = params.preview,
@@ -159,11 +152,9 @@ private fun PreviewCompact(
             onClick = { },
             onLongClick = { },
             onLeadingClick = { },
-            onFavouriteChange = { },
             modifier = Modifier.padding(MainTheme.spacings.double),
             hasAttachments = params.hasAttachments,
             selected = params.selected,
-            favourite = params.favourite,
             threadCount = params.threadCount,
             swapSenderWithSubject = params.swapSenderWithSubject,
             contentPadding = MessageItemDefaults.compactContentPadding,
@@ -176,10 +167,10 @@ private fun PreviewCompact(
 @Preview
 @Composable
 private fun PreviewRelaxed(
-    @PreviewParameter(ReadMessageItemPrevParamCol::class) params: MessageItemPrevParams,
+    @PreviewParameter(JunkMessageItemPrevParamCol::class) params: MessageItemPrevParams,
 ) {
     PreviewWithThemes {
-        ReadMessageItem(
+        JunkMessageItem(
             sender = params.sender,
             subject = params.subject,
             preview = params.preview,
@@ -188,11 +179,9 @@ private fun PreviewRelaxed(
             onClick = { },
             onLongClick = { },
             onLeadingClick = { },
-            onFavouriteChange = { },
             modifier = Modifier.padding(MainTheme.spacings.double),
             hasAttachments = params.hasAttachments,
             selected = params.selected,
-            favourite = params.favourite,
             threadCount = params.threadCount,
             swapSenderWithSubject = params.swapSenderWithSubject,
             contentPadding = MessageItemDefaults.relaxedContentPadding,
@@ -205,10 +194,10 @@ private fun PreviewRelaxed(
 @Preview
 @Composable
 private fun PreviewDefaultWithoutAccountIndicator(
-    @PreviewParameter(ReadMessageItemPrevParamCol::class) params: MessageItemPrevParams,
+    @PreviewParameter(JunkMessageItemPrevParamCol::class) params: MessageItemPrevParams,
 ) {
     PreviewWithThemes {
-        ReadMessageItem(
+        JunkMessageItem(
             sender = params.sender,
             subject = params.subject,
             preview = params.preview,
@@ -229,11 +218,9 @@ private fun PreviewDefaultWithoutAccountIndicator(
             onClick = { },
             onLongClick = { },
             onLeadingClick = { },
-            onFavouriteChange = { },
             modifier = Modifier.padding(MainTheme.spacings.double),
             hasAttachments = params.hasAttachments,
             selected = params.selected,
-            favourite = params.favourite,
             threadCount = params.threadCount,
             swapSenderWithSubject = params.swapSenderWithSubject,
             showAccountIndicator = false,
