+++ feature/mail/message/list/api/src/main/kotlin/net/thunderbird/feature/mail/message/list/ui/component/organism/NewMessageItem.kt
@@ -1,18 +1,24 @@
-package net.thunderbird.core.ui.compose.designsystem.organism.message
+package net.thunderbird.feature.mail.message.list.ui.component.organism
 
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.style.TextOverflow
-import app.k9mail.core.ui.compose.designsystem.atom.text.TextBodyMedium
+import app.k9mail.core.ui.compose.designsystem.atom.text.TextLabelLarge
 import app.k9mail.core.ui.compose.theme2.MainTheme
 import net.thunderbird.core.ui.compose.designsystem.atom.button.FavouriteButtonIcon
-import net.thunderbird.core.ui.compose.designsystem.molecule.message.MessageItemSenderBodyMedium
+import net.thunderbird.core.ui.compose.designsystem.atom.icon.BadgeIcon
+import net.thunderbird.core.ui.compose.designsystem.atom.icon.BadgeIcons
+import net.thunderbird.feature.mail.message.list.ui.component.molecule.MessageItemSenderTitleSmall
+
+private const val NEW_MAIL_BADGE_COLOR = 0xFFF4C430
 
 /**
- * Represents a message item in its Active state.
+ * Represents a message item in its New Message state.
  *
  * @param sender The name of the sender.
  * @param subject The subject of the message.
@@ -37,7 +43,7 @@ import net.thunderbird.core.ui.compose.designsystem.molecule.message.MessageItem
  */
 @Suppress("LongParameterList")
 @Composable
-fun ActiveMessageItem(
+fun NewMessageItem(
     sender: String,
     subject: String,
     preview: String,
@@ -58,7 +64,7 @@ fun ActiveMessageItem(
     contentPadding: PaddingValues = MessageItemDefaults.defaultContentPadding,
     swapSenderWithSubject: Boolean = false,
 ) {
-    ActiveMessageItem(
+    NewMessageItem(
         sender = AnnotatedString(sender),
         subject = subject,
         preview = preview,
@@ -82,7 +88,7 @@ fun ActiveMessageItem(
 }
 
 /**
- * Represents a message item in its Active state.
+ * Represents a message item in its New Message state.
  *
  * @param sender The name of the sender.
  * @param subject The subject of the message.
@@ -107,7 +113,7 @@ fun ActiveMessageItem(
  */
 @Suppress("LongParameterList")
 @Composable
-fun ActiveMessageItem(
+fun NewMessageItem(
     sender: AnnotatedString,
     subject: String,
     preview: String,
@@ -129,21 +135,31 @@ fun ActiveMessageItem(
     swapSenderWithSubject: Boolean = false,
 ) {
     MessageItem(
-        leading = avatar,
+        leading = {
+            Box {
+                avatar()
+                BadgeIcon(
+                    imageVector = BadgeIcons.Filled.NewMail,
+                    tint = Color(NEW_MAIL_BADGE_COLOR),
+                    contentDescription = null,
+                    modifier = Modifier.padding(start = MainTheme.spacings.quarter, top = MainTheme.spacings.quarter),
+                )
+            }
+        },
         sender = {
-            MessageItemSenderBodyMedium(
+            MessageItemSenderTitleSmall(
                 sender = sender,
                 subject = subject,
                 swapSenderWithSubject = swapSenderWithSubject,
                 threadCount = threadCount,
-                color = MainTheme.colors.onSurfaceVariant,
+                color = if (swapSenderWithSubject) MainTheme.colors.primary else MainTheme.colors.onSurface,
             )
         },
         subject = {
             if (swapSenderWithSubject) {
-                TextBodyMedium(text = sender, maxLines = 1, overflow = TextOverflow.Ellipsis)
+                TextLabelLarge(text = sender, maxLines = 1, overflow = TextOverflow.Ellipsis)
             } else {
-                TextBodyMedium(text = subject, maxLines = 1, overflow = TextOverflow.Ellipsis)
+                TextLabelLarge(text = subject, maxLines = 1, overflow = TextOverflow.Ellipsis)
             }
         },
         preview = preview,
@@ -157,7 +173,9 @@ fun ActiveMessageItem(
         colors = if (selected) {
             MessageItemDefaults.selectedMessageItemColors()
         } else {
-            MessageItemDefaults.activeMessageItemColors()
+            MessageItemDefaults.newMessageItemColors(
+                subjectColor = if (swapSenderWithSubject) MainTheme.colors.onSurface else MainTheme.colors.primary,
+            )
         },
         modifier = modifier,
         hasAttachments = hasAttachments,
